Add logic to the index view for the cows resource.
Why:

In order to ensure that our cows#index action is properly "wired" with
the index view, we have display an item in the index view from the
cows#index action.

This change addresses the need by:

Creating logic in the index view that iterates over a collection of
cows, and displays each cow's details.