Add a create action in the Cows controller for handling a user's request to create a new cow.
Why:

It is necessary to define a create action to receive a post request for
creating a new resource.

This change addresses the need by:

Creates a create action in the Cows controller. Note that in that
resource's controller, we also define a private method that whitelists
which parameters can be mass-assignable.